Add missing scripts/ directory and contents
This commit does not exist in the actual SVN mirror. Sometime after revision r8641, the scripts/ directory was added to the moos-ivp-aro repo (whereas before it was only accessible in the private tree).

In SVN this is acceptable, but in git it effectively rewrites history (the scripts/ directory existed back in r2407). To avoid rewriting history again here, we will add the scripts directory as of its 'most recent' state in r8641.

The following line is _fake news_, and tricks git-svn into thinking this commit is in SVN history, therefore re-enabling `git svn fetch`. There may be a better way to do this, but I'm not aware of it.
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@8670 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf